split TileCache into several pieces
